Add files via upload
I am aware that this is not the exact same as most of my classmates, I did not have any of the lab code from the class sessions, so I just went off the directions you gave and also the images you provided. I know that the rectangles might not be the exact size, and I also noticed that in the background was black in your example. I kept mine red because the directions never said the change it to black. But just in case, I am aware that all I would have to do was changed the rgb values to (0,0,0).